Move auxilliary data to separate module.
